+++ test/cli/silent/_config.js
@@ -1,0 +1,5 @@
+module.exports = {
+	description: 'does not print warnings with --silent',
+	command: 'rollup -i main.js -f cjs --silent',
+	stderr: ``
+};