+++ test/form/class-constructor-side-effect/_config.js
@@ -1,0 +1,3 @@
+module.exports = {
+	description: 'preserves side-effects in a class constructor (#1233)'
+};