+++ test/form/class-constructor-side-effect/main.js
@@ -1,0 +1,7 @@
+class Foo {
+	constructor () {
+		console.log( 'Foo' );
+	}
+}
+
+new Foo;